Run additional tests on my_methods
